Embed YT videos
I thought it could be better if the watchers could have only one link for the videos and materials. 

There are two ways to embed YT vide; there is the one-liner option with knitr::include_url, but no fullscreen is possible with this option (AFAIK).

To enable fullscreen videos, I had to add some CSS classes and include the iframe within a *div container*. This way, the iframe takes 100% of the width, and the height gets adjusted proportionally. 

 I included both solutions (in `index.Rmd`) in case you prefer one over the other. As a watcher, I feel the second one with full-screen is more enjoyable.